+++ src/routes/auth/change-email/direct-change.ts
@@ -1,12 +1,13 @@
-import { Request, Response } from 'express'
+import { Response } from 'express'
 
 import { asyncWrapper } from '@shared/helpers'
 import { changeEmailByUserId } from '@shared/queries'
 import { request } from '@shared/request'
 
 import { getRequestInfo } from './utils'
+import { RequestExtended } from '@shared/types'
 
-async function requestChangeEmail(req: Request, res: Response): Promise<unknown> {
+async function requestChangeEmail(req: RequestExtended, res: Response): Promise<unknown> {
   const { user_id, new_email } = await getRequestInfo(req)
 
   // * Email verification is not activated - change email straight away