+++ src/routes/auth/change-password/reset.ts
@@ -1,4 +1,4 @@
-import { Request, Response } from 'express'
+import { Response } from 'express'
 import { v4 as uuidv4 } from 'uuid'
 import Boom from '@hapi/boom'
 
@@ -6,12 +6,12 @@ import { asyncWrapper, checkHibp, hashPassword } from '@shared/helpers'
 import { resetPasswordWithTicketSchema } from '@shared/validation'
 import { updatePasswordWithTicket } from '@shared/queries'
 import { request } from '@shared/request'
-import { UpdateAccountData } from '@shared/types'
+import { UpdateAccountData, RequestExtended } from '@shared/types'
 
 /**
  * Reset the password, either from a valid ticket or from a valid JWT and a valid password
  */
-async function resetPassword(req: Request, res: Response): Promise<unknown> {
+async function resetPassword(req: RequestExtended, res: Response): Promise<unknown> {
   // Reset the password from { ticket, new_password }
   const { ticket, new_password } = await resetPasswordWithTicketSchema.validateAsync(req.body)