+++ src/routes/auth/jwks.ts
@@ -1,9 +1,10 @@
-import { Request, Response } from 'express'
+import { Response } from 'express'
 
 import { asyncWrapper } from '@shared/helpers'
 import { getJwkStore } from '@shared/jwt'
+import { RequestExtended } from '@shared/types'
 
-const getJwks = async (_req: Request, res: Response): Promise<Response<unknown>> =>
+const getJwks = async (_req: RequestExtended, res: Response): Promise<Response<unknown>> =>
   res.send(getJwkStore().toJWKS(false))
 
 export default asyncWrapper(getJwks)