+++ src/routes/auth/token/refresh.ts
@@ -1,29 +1,30 @@
 import { asyncWrapper } from '@shared/helpers'
-import { Request, Response } from 'express'
+import { Response } from 'express'
 import { selectRefreshToken, updateRefreshToken } from '@shared/queries'
 
 import Boom from '@hapi/boom'
-import { COOKIE_SECRET } from '@shared/config'
 import { newJwtExpiry, createHasuraJwt, generatePermissionVariables } from '@shared/jwt'
 import { newRefreshExpiry, setCookie } from '@shared/cookies'
 import { request } from '@shared/request'
 import { v4 as uuidv4 } from 'uuid'
-import { AccountData } from '@shared/types'
+import { AccountData, RequestExtended } from '@shared/types'
 
 interface HasuraData {
   auth_refresh_tokens: { account: AccountData }[]
 }
 
-async function refreshToken({ cookies, signedCookies }: Request, res: Response): Promise<unknown> {
-  const { refresh_token } = COOKIE_SECRET ? signedCookies : cookies
-
+async function refreshToken({ refresh_token }: RequestExtended, res: Response): Promise<void> {
   if (!refresh_token) {
+    throw Boom.unauthorized('Invalid or expired refresh token.')
+  }
+
+  if (!refresh_token.value) {
     throw Boom.unauthorized('Invalid or expired refresh token.')
   }
 
   // get account based on refresh token
   const { auth_refresh_tokens } = await request<HasuraData>(selectRefreshToken, {
-    refresh_token,
+    refresh_token: refresh_token.value,
     current_timestamp: new Date()
   })
 
@@ -39,7 +40,7 @@ async function refreshToken({ cookies, signedCookies }: Request, res: Response):
   // and insert new refresh token
   try {
     await request(updateRefreshToken, {
-      old_refresh_token: refresh_token,
+      old_refresh_token: refresh_token.value,
       new_refresh_token_data: {
         account_id: account.id,
         refresh_token: new_refresh_token,
@@ -52,12 +53,22 @@ async function refreshToken({ cookies, signedCookies }: Request, res: Response):
 
   const permission_variables = JSON.stringify(generatePermissionVariables(account))
 
-  setCookie(res, new_refresh_token, permission_variables)
+  const jwt_token = createHasuraJwt(account)
+  const jwt_expires_in = newJwtExpiry
 
-  return res.send({
-    jwt_token: createHasuraJwt(account),
-    jwt_expires_in: newJwtExpiry
-  })
+  if (refresh_token.type === 'cookie') {
+    setCookie(res, new_refresh_token, permission_variables)
+    res.send({
+      jwt_token,
+      jwt_expires_in
+    })
+  } else {
+    res.send({
+      jwt_token,
+      jwt_expires_in,
+      refresh_token: new_refresh_token
+    })
+  }
 }
 
 export default asyncWrapper(refreshToken)