+++ src/routes/storage/delete.ts
@@ -1,4 +1,4 @@
-import { NextFunction, Request, Response } from 'express'
+import { NextFunction, Response } from 'express'
 import {
   PathConfig,
   createContext,
@@ -11,9 +11,10 @@ import {
 import Boom from '@hapi/boom'
 import { S3_BUCKET } from '@shared/config'
 import { s3 } from '@shared/s3'
+import { RequestExtended } from '@shared/types'
 
 export const deleteFile = async (
-  req: Request,
+  req: RequestExtended,
   res: Response,
   _next: NextFunction,
   rules: Partial<PathConfig>,