+++ src/routes/storage/upload.ts
@@ -1,4 +1,4 @@
-import { NextFunction, Request, Response } from 'express'
+import { NextFunction, Response } from 'express'
 import { v4 as uuidv4 } from 'uuid'
 import { PathConfig, createContext, getHeadObject, getKey, hasPermission } from './utils'
 
@@ -6,9 +6,10 @@ import Boom from '@hapi/boom'
 import { S3_BUCKET } from '@shared/config'
 import { UploadedFile } from 'express-fileupload'
 import { s3 } from '@shared/s3'
+import { RequestExtended } from '@shared/types'
 
 export const uploadFile = async (
-  req: Request,
+  req: RequestExtended,
   res: Response,
   _next: NextFunction,
   rules: Partial<PathConfig>,