+++ src/routes/storage/utils.ts
@@ -3,14 +3,12 @@ import { v4 as uuidv4 } from 'uuid'
 
 import Boom from '@hapi/boom'
 import { HeadObjectOutput } from 'aws-sdk/clients/s3'
-import { Request } from 'express'
 import { S3_BUCKET } from '@shared/config'
 import fs from 'fs'
 import path from 'path'
 import { s3 } from '@shared/s3'
 import yaml from 'js-yaml'
-import { PermissionVariables } from '@shared/types'
-import { getPermissionVariablesFromCookie } from '@shared/helpers'
+import { PermissionVariables, RequestExtended } from '@shared/types'
 
 export const OBJECT_PREFIX = '/o'
 export const META_PREFIX = '/m'
@@ -81,15 +79,10 @@ const storageFunctions = (context: object): { [key: string]: Function } =>
   }, {})
 
 export const createContext = (
-  req: Request,
+  req: RequestExtended,
   s3HeadObject: object = {} // TODO better s3 head object type
 ): object => {
-  let auth
-  try {
-    auth = getPermissionVariablesFromCookie(req)
-  } catch (err) {
-    auth = undefined
-  }
+  const auth = req.permission_variables
 
   const variables: StorageContext = {
     request: {
@@ -125,10 +118,10 @@ export const generateMetadata = (metadataParams: object, context: object): objec
   }, {})
 
 // Creates an object key that is the path without the first character '/'
-export const getKey = (req: Request): string => req.path.substring(1)
+export const getKey = (req: RequestExtended): string => req.path.substring(1)
 
 export const getHeadObject = async (
-  req: Request,
+  req: RequestExtended,
   ignoreErrors = false
 ): Promise<HeadObjectOutput | undefined> => {
   const params = {
@@ -146,7 +139,7 @@ export const getHeadObject = async (
 }
 
 export const replaceMetadata = async (
-  req: Request,
+  req: RequestExtended,
   keepOldMetadata: boolean,
   newMetadata: object = {}
 ): Promise<void> => {