+++ src/shared/types.ts
@@ -1,3 +1,5 @@
+import { Request } from 'express'
+
 export type ClaimValueType =
   | string
   | string[]
@@ -89,6 +91,16 @@ export interface InsertAccountData {
   insert_auth_accounts: {
     returning: AccountData[]
   }
+}
+
+export interface RefreshTokenMiddleware {
+  value: string | null
+  type: 'query' | 'cookie' | null
+}
+
+export interface RequestExtended extends Request {
+  refresh_token?: RefreshTokenMiddleware
+  permission_variables?: PermissionVariables
 }
 
 export interface SetNewEmailData {